Add Home Page
From here we need to build and navigate to the
account pages.

This was a very challenging page to build. All
other pages in the app are static, at least on the
UI level. This makes them fairly straightforward
to build and let the model update them and the
controller coordinate it all.

The number of information panels on this page is not
fixed however and must account for unknown info at
the time of building.