Add statement page layout
The page loads off sample data. Next we will need
to wire up the db queries to populate data.

This step introduced a major change to the database.
We had to add a Balance column to account activity
in order to track historical balances. This itself
is a simple add to the DB but represents a major
task in terms of populating sample data with
accurate test data.

Once populated, this column should let us go back
and refactor some code where we used math intensive
algorithms to fill in balance data.