Add Transactions
Users can deposit, withdraw, or transfer funds.
Payments are not yet installed. Success screen is
not yet implemented. This has a temp bug of being
able to commit multiple transactions in a single
page which will not properly update balances. This
is expected to be resolved once the navigation is
installed.

Adding new payees is not yet installed.