Add Insert New Payees
Entering a new payee in the transaction page will
now update the DB. It is possible for a user  to
key in a description that exists in the db but not
linked to their account. In this case, the app first
searches for possible matches and will return an
existing ID if found.